Fix multi-rigitbody-crash in NatNet >= v2.6
Problem was caused by modification of data structure. Details can be seen in CrashFix.md.

This commit mainly added some code to handle the extra track_flag data
field, and keep compatible with old code.

The track_flag is parsed and stored. The publisher will check this
variable, and skip publishing if the rigidbody is not published.

Read version information from yaml. It is treated as a float.